update an analyze result tool
